backend/requirements.txt
Flask
Flask-CORS
pytesseract
pillow
